requirements.txt
aiogram==2.13
gunicorn==20.0.4
aiohttp==3.7.4
asgiref==3.3.4
python-dotenv==0.17.1
psycopg2-binary==2.9.1
alembic==1.6.5
aniso8601==9.0.1
appdirs==1.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
cachetools==4.2.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
fire==0.4.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
oauthlib==3.1.1
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.10
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2020.5
pyyaxml==0.6.5
reportlab==3.5.68
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.4.19
sqlparse==0.4.1
termcolor==1.1.0
traitlets==5.0.5
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
xmltodict==0.12.0
yarl==1.6.3
sqlalchemy_serializer
